+++ packages/terra-clinical-onset-picker/src/terra-dev-site/doc/example/DefaultOnset.jsx
@@ -9,6 +9,7 @@ const picker = () => (
   <div>
     <p>{`Birthdate: ${birthdate.format('MM-DD-YYYY')}`}</p>
     <OnsetPicker
+      legend="Onset of symptoms"
       birthdate={birthdate.format('YYYY-MM-DD')}
       id="doogs"
     />