+++ packages/terra-clinical-onset-picker/src/terra-dev-site/doc/example/HandledOnset.jsx
@@ -57,6 +57,7 @@ class HandledOnsetExample extends React.Component {
             'on/at',
           ]}
           onsetDate={this.state.onsetDate}
+          legend="Date of Birth"
           onsetOnChange={this.handleOnset}
         />
         <button type="submit">Submit</button>