+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Added
 * Documentation of violation scenarios.
 * Indicate in site examples what birthdate is set to.
+* Added legend and isLegendHidden prop to Onsetpicker field group.
 
 4.2.0 - (April 16, 2019)
 ----------